Ensure a broken build if controllers exist with ambiguous constructors.
The previous approach made the Web.UI project needlessly different to
a downstream project created by using Umbraco.Templates.

With this setup, any controllers with ambiguous constructors will result
in all TestServerTestBase tests failing (because these verify
the container in the base class).

To see this in action, comment out the call to
AddControllersWithAmbiguousConstructors (AddBackOfficeCore)
and run UserControllerTests.
